Made framework more configurable by removing constants and gce-pd specgcepd specific code into variables
